Struct aws_sdk_cloudwatch::client::fluent_builders::GetMetricData

pub struct GetMetricData<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> GetMetricData<C>

pub async fn send(
self
) -> Result<GetMetricDataOutput, SdkError<GetMetricDataError>> where
C: SmithyConnector,

pub fn metric_data_queries(self, inp: impl Into<MetricDataQuery>) -> Self

The metric queries to be returned. A single GetMetricData call can include as many as 500 MetricDataQuery structures. Each of these structures can specify either a metric to retrieve, or a math expression to perform on retrieved data.

pub fn set_metric_data_queries(
self,
input: Option<Vec<MetricDataQuery>>
) -> Self

pub fn start_time(self, input: Instant) -> Self

The time stamp indicating the earliest data to be returned.

The value specified is inclusive; results include data points with the specified time stamp.

CloudWatch rounds the specified time stamp as follows:

Start time less than 15 days ago - Round down to the nearest whole minute. For example, 12:32:34 is rounded down to 12:32:00.

Start time between 15 and 63 days ago - Round down to the nearest 5-minute clock interval. For example, 12:32:34 is rounded down to 12:30:00.

Start time greater than 63 days ago - Round down to the nearest 1-hour clock interval. For example, 12:32:34 is rounded down to 12:00:00.

If you set Period to 5, 10, or 30, the start time of your request is rounded down to the nearest time that corresponds to even 5-, 10-, or 30-second divisions of a minute. For example, if you make a query at (HH:mm:ss) 01:05:23 for the previous 10-second period, the start time of your request is rounded down and you receive data from 01:05:10 to 01:05:20. If you make a query at 15:07:17 for the previous 5 minutes of data, using a period of 5 seconds, you receive data timestamped between 15:02:15 and 15:07:15.

For better performance, specify StartTime and EndTime values that align with the value of the metric's Period and sync up with the beginning and end of an hour. For example, if the Period of a metric is 5 minutes, specifying 12:05 or 12:30 as StartTime can get a faster response from CloudWatch than setting 12:07 or 12:29 as the StartTime.

pub fn set_start_time(self, input: Option<Instant>) -> Self

pub fn end_time(self, input: Instant) -> Self

The time stamp indicating the latest data to be returned.

The value specified is exclusive; results include data points up to the specified time stamp.

For better performance, specify StartTime and EndTime values that align with the value of the metric's Period and sync up with the beginning and end of an hour. For example, if the Period of a metric is 5 minutes, specifying 12:05 or 12:30 as EndTime can get a faster response from CloudWatch than setting 12:07 or 12:29 as the EndTime.

pub fn set_end_time(self, input: Option<Instant>) -> Self

pub fn next_token(self, input: impl Into<String>) -> Self

Include this value, if it was returned by the previous GetMetricData operation, to get the next set of data points.

pub fn set_next_token(self, input: Option<String>) -> Self

pub fn scan_by(self, input: ScanBy) -> Self

The order in which data points should be returned. TimestampDescending returns the newest data first and paginates when the MaxDatapoints limit is reached. TimestampAscending returns the oldest data first and paginates when the MaxDatapoints limit is reached.

pub fn set_scan_by(self, input: Option<ScanBy>) -> Self

pub fn max_datapoints(self, input: i32) -> Self

The maximum number of data points the request should return before paginating. If you omit this, the default of 100,800 is used.

pub fn set_max_datapoints(self, input: Option<i32>) -> Self

pub fn label_options(self, input: LabelOptions) -> Self

This structure includes the Timezone parameter, which you can use to specify your time zone so that the labels of returned data display the correct time for your time zone.

pub fn set_label_options(self, input: Option<LabelOptions>) -> Self