Struct aws_sdk_cloudwatch::client::fluent_builders::GetMetricWidgetImage

pub struct GetMetricWidgetImage<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> GetMetricWidgetImage<C>

pub async fn send(
self
) -> Result<GetMetricWidgetImageOutput, SdkError<GetMetricWidgetImageError>> where
C: SmithyConnector,

pub fn metric_widget(self, input: impl Into<String>) -> Self

A JSON string that defines the bitmap graph to be retrieved. The string includes the metrics to include in the graph, statistics, annotations, title, axis limits, and so on. You can include only one MetricWidget parameter in each GetMetricWidgetImage call.

For more information about the syntax of MetricWidget see GetMetricWidgetImage: Metric Widget Structure and Syntax.

If any metric on the graph could not load all the requested data points, an orange triangle with an exclamation point appears next to the graph legend.

pub fn set_metric_widget(self, input: Option<String>) -> Self

pub fn output_format(self, input: impl Into<String>) -> Self

The format of the resulting image. Only PNG images are supported.

The default is png. If you specify png, the API returns an HTTP response with the content-type set to text/xml. The image data is in a MetricWidgetImage field. For example:

>

iVBORw0KGgoAAAANSUhEUgAAAlgAAAGQEAYAAAAip...

6f0d4192-4d42-11e8-82c1-f539a07e0e3b

The image/png setting is intended only for custom HTTP requests. For most use cases, and all actions using an AWS SDK, you should use png. If you specify image/png, the HTTP response has a content-type set to image/png, and the body of the response is a PNG image.

pub fn set_output_format(self, input: Option<String>) -> Self