Struct aws_sdk_cloudwatch::client::fluent_builders::PutCompositeAlarm

pub struct PutCompositeAlarm<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> PutCompositeAlarm<C>

pub async fn send(
self
) -> Result<PutCompositeAlarmOutput, SdkError<PutCompositeAlarmError>> where
C: SmithyConnector,

pub fn actions_enabled(self, input: bool) -> Self

Indicates whether actions should be executed during any changes to the alarm state of the composite alarm. The default is TRUE.

pub fn set_actions_enabled(self, input: Option<bool>) -> Self

pub fn alarm_actions(self, inp: impl Into<String>) -> Self

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:sns:region:account-id:sns-topic-name | arn:aws:ssm:region:account-id:opsitem:severity

pub fn set_alarm_actions(self, input: Option<Vec<String>>) -> Self

pub fn alarm_description(self, input: impl Into<String>) -> Self

The description for the composite alarm.

pub fn set_alarm_description(self, input: Option<String>) -> Self

pub fn alarm_name(self, input: impl Into<String>) -> Self

The name for the composite alarm. This name must be unique within the Region.

pub fn set_alarm_name(self, input: Option<String>) -> Self

pub fn alarm_rule(self, input: impl Into<String>) -> Self

An expression that specifies which other alarms are to be evaluated to determine this composite alarm's state. For each alarm that you reference, you designate a function that specifies whether that alarm needs to be in ALARM state, OK state, or INSUFFICIENT_DATA state. You can use operators (AND, OR and NOT) to combine multiple functions in a single expression. You can use parenthesis to logically group the functions in your expression.

You can use either alarm names or ARNs to reference the other alarms that are to be evaluated.

Functions can include the following:

ALARM("alarm-name or alarm-ARN") is TRUE if the named alarm is in ALARM state.

OK("alarm-name or alarm-ARN") is TRUE if the named alarm is in OK state.

INSUFFICIENT_DATA("alarm-name or alarm-ARN") is TRUE if the named alarm is in INSUFFICIENT_DATA state.

TRUE always evaluates to TRUE.

FALSE always evaluates to FALSE.

TRUE and FALSE are useful for testing a complex AlarmRule structure, and for testing your alarm actions.

Alarm names specified in AlarmRule can be surrounded with double-quotes ("), but do not have to be.

The following are some examples of AlarmRule:

ALARM(CPUUtilizationTooHigh) AND ALARM(DiskReadOpsTooHigh) specifies that the composite alarm goes into ALARM state only if both CPUUtilizationTooHigh and DiskReadOpsTooHigh alarms are in ALARM state.

ALARM(CPUUtilizationTooHigh) AND NOT ALARM(DeploymentInProgress) specifies that the alarm goes to ALARM state if CPUUtilizationTooHigh is in ALARM state and DeploymentInProgress is not in ALARM state. This example reduces alarm noise during a known deployment window.

(ALARM(CPUUtilizationTooHigh) OR ALARM(DiskReadOpsTooHigh)) AND OK(NetworkOutTooHigh) goes into ALARM state if CPUUtilizationTooHigh OR DiskReadOpsTooHigh is in ALARM state, and if NetworkOutTooHigh is in OK state. This provides another example of using a composite alarm to prevent noise. This rule ensures that you are not notified with an alarm action on high CPU or disk usage if a known network problem is also occurring.

The AlarmRule can specify as many as 100 "children" alarms. The AlarmRule expression can have as many as 500 elements. Elements are child alarms, TRUE or FALSE statements, and parentheses.

pub fn set_alarm_rule(self, input: Option<String>) -> Self

pub fn insufficient_data_actions(self, inp: impl Into<String>) -> Self

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:sns:region:account-id:sns-topic-name

pub fn set_insufficient_data_actions(self, input: Option<Vec<String>>) -> Self

pub fn ok_actions(self, inp: impl Into<String>) -> Self

The actions to execute when this alarm transitions to an OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:sns:region:account-id:sns-topic-name

pub fn set_ok_actions(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

A list of key-value pairs to associate with the composite alarm. You can associate as many as 50 tags with an alarm.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions, by granting a user permission to access or change only resources with certain tag values.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self