Module aws_sdk_cloudwatch::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

dashboard_invalid_input_error

See DashboardInvalidInputError

dashboard_not_found_error

See DashboardNotFoundError

internal_service_fault

See InternalServiceFault

invalid_format_fault

See InvalidFormatFault

invalid_next_token

See InvalidNextToken

invalid_parameter_combination_exception

See InvalidParameterCombinationException

invalid_parameter_value_exception

See InvalidParameterValueException

limit_exceeded_exception

See LimitExceededException

limit_exceeded_fault

See LimitExceededFault

missing_required_parameter_exception

See MissingRequiredParameterException

resource_not_found

See ResourceNotFound

resource_not_found_exception

See ResourceNotFoundException

Structs

ConcurrentModificationException

More than one process tried to modify a resource at the same time.

DashboardInvalidInputError

Some part of the dashboard data is invalid.

DashboardNotFoundError

The specified dashboard does not exist.

DeleteAlarmsError

DeleteAnomalyDetectorError

DeleteDashboardsError

DeleteInsightRulesError

DeleteMetricStreamError

DescribeAlarmHistoryError

DescribeAlarmsError

DescribeAlarmsForMetricError

DescribeAnomalyDetectorsError

DescribeInsightRulesError

DisableAlarmActionsError

DisableInsightRulesError

EnableAlarmActionsError

EnableInsightRulesError

GetDashboardError

GetInsightRuleReportError

GetMetricDataError

GetMetricStatisticsError

GetMetricStreamError

GetMetricWidgetImageError

InternalServiceFault

Request processing has failed due to some unknown error, exception, or failure.

InvalidFormatFault

Data was not syntactically valid JSON.

InvalidNextToken

The next token specified is invalid.

InvalidParameterCombinationException

Parameters were used together that cannot be used together.

InvalidParameterValueException

The value of an input parameter is bad or out-of-range.

LimitExceededException

The operation exceeded one or more limits.

LimitExceededFault

The quota for alarms for this customer has already been reached.

ListDashboardsError

ListMetricStreamsError

ListMetricsError

ListTagsForResourceError

MissingRequiredParameterException

An input parameter that is required is missing.

PutAnomalyDetectorError

PutCompositeAlarmError

PutDashboardError

PutInsightRuleError

PutMetricAlarmError

PutMetricDataError

PutMetricStreamError

ResourceNotFound

The named resource does not exist.

ResourceNotFoundException

The named resource does not exist.

SetAlarmStateError

StartMetricStreamsError

StopMetricStreamsError

TagResourceError

UntagResourceError

Enums

DeleteAlarmsErrorKind

DeleteAnomalyDetectorErrorKind

DeleteDashboardsErrorKind

DeleteInsightRulesErrorKind

DeleteMetricStreamErrorKind

DescribeAlarmHistoryErrorKind

DescribeAlarmsErrorKind

DescribeAlarmsForMetricErrorKind

DescribeAnomalyDetectorsErrorKind

DescribeInsightRulesErrorKind

DisableAlarmActionsErrorKind

DisableInsightRulesErrorKind

EnableAlarmActionsErrorKind

EnableInsightRulesErrorKind

GetDashboardErrorKind

GetInsightRuleReportErrorKind

GetMetricDataErrorKind

GetMetricStatisticsErrorKind

GetMetricStreamErrorKind

GetMetricWidgetImageErrorKind

ListDashboardsErrorKind

ListMetricStreamsErrorKind

ListMetricsErrorKind

ListTagsForResourceErrorKind

PutAnomalyDetectorErrorKind

PutCompositeAlarmErrorKind

PutDashboardErrorKind

PutInsightRuleErrorKind

PutMetricAlarmErrorKind

PutMetricDataErrorKind

PutMetricStreamErrorKind

SetAlarmStateErrorKind

StartMetricStreamsErrorKind

StopMetricStreamsErrorKind

TagResourceErrorKind

UntagResourceErrorKind