Struct aws_sdk_cloudwatch::input::put_metric_alarm_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for PutMetricAlarmInput

Implementations

impl Builder

pub fn alarm_name(self, input: impl Into<String>) -> Self

The name for the alarm. This name must be unique within the Region.

pub fn set_alarm_name(self, input: Option<String>) -> Self

pub fn alarm_description(self, input: impl Into<String>) -> Self

The description for the alarm.

pub fn set_alarm_description(self, input: Option<String>) -> Self

pub fn actions_enabled(self, input: bool) -> Self

Indicates whether actions should be executed during any changes to the alarm state. The default is TRUE.

pub fn set_actions_enabled(self, input: Option<bool>) -> Self

pub fn ok_actions(self, input: impl Into<String>) -> Self

pub fn set_ok_actions(self, input: Option<Vec<String>>) -> Self

pub fn alarm_actions(self, input: impl Into<String>) -> Self

pub fn set_alarm_actions(self, input: Option<Vec<String>>) -> Self

pub fn insufficient_data_actions(self, input: impl Into<String>) -> Self

pub fn set_insufficient_data_actions(self, input: Option<Vec<String>>) -> Self

pub fn metric_name(self, input: impl Into<String>) -> Self

The name for the metric associated with the alarm. For each PutMetricAlarm operation, you must specify either MetricName or a Metrics array.

If you are creating an alarm based on a math expression, you cannot specify this parameter, or any of the Dimensions, Period, Namespace, Statistic, or ExtendedStatistic parameters. Instead, you specify all this information in the Metrics array.

pub fn set_metric_name(self, input: Option<String>) -> Self

pub fn namespace(self, input: impl Into<String>) -> Self

The namespace for the metric associated specified in MetricName.

pub fn set_namespace(self, input: Option<String>) -> Self

pub fn statistic(self, input: Statistic) -> Self

The statistic for the metric specified in MetricName, other than percentile. For percentile statistics, use ExtendedStatistic. When you call PutMetricAlarm and specify a MetricName, you must specify either Statistic or ExtendedStatistic, but not both.

pub fn set_statistic(self, input: Option<Statistic>) -> Self

pub fn extended_statistic(self, input: impl Into<String>) -> Self

The percentile statistic for the metric specified in MetricName. Specify a value between p0.0 and p100. When you call PutMetricAlarm and specify a MetricName, you must specify either Statistic or ExtendedStatistic, but not both.

pub fn set_extended_statistic(self, input: Option<String>) -> Self

pub fn dimensions(self, input: impl Into<Dimension>) -> Self

pub fn set_dimensions(self, input: Option<Vec<Dimension>>) -> Self

pub fn period(self, input: i32) -> Self

The length, in seconds, used each time the metric specified in MetricName is evaluated. Valid values are 10, 30, and any multiple of 60.

Period is required for alarms based on static thresholds. If you are creating an alarm based on a metric math expression, you specify the period for each metric within the objects in the Metrics array.

Be sure to specify 10 or 30 only for metrics that are stored by a PutMetricData call with a StorageResolution of 1. If you specify a period of 10 or 30 for a metric that does not have sub-minute resolution, the alarm still attempts to gather data at the period rate that you specify. In this case, it does not receive data for the attempts that do not correspond to a one-minute data resolution, and the alarm might often lapse into INSUFFICENT_DATA status. Specifying 10 or 30 also sets this alarm as a high-resolution alarm, which has a higher charge than other alarms. For more information about pricing, see Amazon CloudWatch Pricing.

An alarm's total current evaluation period can be no longer than one day, so Period multiplied by EvaluationPeriods cannot be more than 86,400 seconds.

pub fn set_period(self, input: Option<i32>) -> Self

pub fn unit(self, input: StandardUnit) -> Self

The unit of measure for the statistic. For example, the units for the Amazon EC2 NetworkIn metric are Bytes because NetworkIn tracks the number of bytes that an instance receives on all network interfaces. You can also specify a unit when you create a custom metric. Units help provide conceptual meaning to your data. Metric data points that specify a unit of measure, such as Percent, are aggregated separately.

If you don't specify Unit, CloudWatch retrieves all unit types that have been published for the metric and attempts to evaluate the alarm. Usually, metrics are published with only one unit, so the alarm works as intended.

However, if the metric is published with multiple types of units and you don't specify a unit, the alarm's behavior is not defined and it behaves predictably.

We recommend omitting Unit so that you don't inadvertently specify an incorrect unit that is not published for this metric. Doing so causes the alarm to be stuck in the INSUFFICIENT DATA state.

pub fn set_unit(self, input: Option<StandardUnit>) -> Self

pub fn evaluation_periods(self, input: i32) -> Self

The number of periods over which data is compared to the specified threshold. If you are setting an alarm that requires that a number of consecutive data points be breaching to trigger the alarm, this value specifies that number. If you are setting an "M out of N" alarm, this value is the N.

An alarm's total current evaluation period can be no longer than one day, so this number multiplied by Period cannot be more than 86,400 seconds.

pub fn set_evaluation_periods(self, input: Option<i32>) -> Self

pub fn datapoints_to_alarm(self, input: i32) -> Self

The number of data points that must be breaching to trigger the alarm. This is used only if you are setting an "M out of N" alarm. In that case, this value is the M. For more information, see Evaluating an Alarm in the Amazon CloudWatch User Guide.

pub fn set_datapoints_to_alarm(self, input: Option<i32>) -> Self

pub fn threshold(self, input: f64) -> Self

The value against which the specified statistic is compared.

This parameter is required for alarms based on static thresholds, but should not be used for alarms based on anomaly detection models.

pub fn set_threshold(self, input: Option<f64>) -> Self

pub fn comparison_operator(self, input: ComparisonOperator) -> Self

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

The values LessThanLowerOrGreaterThanUpperThreshold, LessThanLowerThreshold, and GreaterThanUpperThreshold are used only for alarms based on anomaly detection models.

pub fn set_comparison_operator(self, input: Option<ComparisonOperator>) -> Self

pub fn treat_missing_data(self, input: impl Into<String>) -> Self

Sets how this alarm is to handle missing data points. If TreatMissingData is omitted, the default behavior of missing is used. For more information, see Configuring How CloudWatch Alarms Treats Missing Data.

Valid Values: breaching | notBreaching | ignore | missing

pub fn set_treat_missing_data(self, input: Option<String>) -> Self

pub fn evaluate_low_sample_count_percentile(
self,
input: impl Into<String>
) -> Self

Used only for alarms based on percentiles. If you specify ignore, the alarm state does not change during periods with too few data points to be statistically significant. If you specify evaluate or omit this parameter, the alarm is always evaluated and possibly changes state no matter how many data points are available. For more information, see Percentile-Based CloudWatch Alarms and Low Data Samples.

Valid Values: evaluate | ignore

pub fn set_evaluate_low_sample_count_percentile(
self,
input: Option<String>
) -> Self

pub fn metrics(self, input: impl Into<MetricDataQuery>) -> Self

pub fn set_metrics(self, input: Option<Vec<MetricDataQuery>>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn threshold_metric_id(self, input: impl Into<String>) -> Self

If this is an alarm based on an anomaly detection model, make this value match the ID of the ANOMALY_DETECTION_BAND function.

For an example of how to use this parameter, see the Anomaly Detection Model Alarm example on this page.

If your alarm uses this parameter, it cannot have Auto Scaling actions.

pub fn set_threshold_metric_id(self, input: Option<String>) -> Self

pub fn build(self) -> Result<PutMetricAlarmInput, BuildError>

Consumes the builder and constructs a PutMetricAlarmInput