Module aws_sdk_cloudwatch::model

Modules

alarm_history_item

See AlarmHistoryItem

anomaly_detector

See AnomalyDetector

anomaly_detector_configuration

See AnomalyDetectorConfiguration

composite_alarm

See CompositeAlarm

dashboard_entry

See DashboardEntry

dashboard_validation_message

See DashboardValidationMessage

datapoint

See Datapoint

dimension

See Dimension

dimension_filter

See DimensionFilter

insight_rule

See InsightRule

insight_rule_contributor

See InsightRuleContributor

insight_rule_contributor_datapoint

See InsightRuleContributorDatapoint

insight_rule_metric_datapoint

See InsightRuleMetricDatapoint

label_options

See LabelOptions

message_data

See MessageData

metric

See Metric

metric_alarm

See MetricAlarm

metric_data_query

See MetricDataQuery

metric_data_result

See MetricDataResult

metric_datum

See MetricDatum

metric_stat

See MetricStat

metric_stream_entry

See MetricStreamEntry

metric_stream_filter

See MetricStreamFilter

partial_failure

See PartialFailure

range

See Range

statistic_set

See StatisticSet

tag

See Tag

Structs

AlarmHistoryItem

Represents the history of a specific alarm.

AnomalyDetector

An anomaly detection model associated with a particular CloudWatch metric and statistic. You

AnomalyDetectorConfiguration

The configuration specifies details about how the anomaly detection model is to be trained,

CompositeAlarm

The details about a composite alarm.

DashboardEntry

Represents a specific dashboard.

DashboardValidationMessage

An error or warning for the operation.

Datapoint

Encapsulates the statistical data that CloudWatch computes from metric data.

Dimension

A dimension is a name/value pair that is part of the identity of a metric. You

DimensionFilter

Represents filters for a dimension.

InsightRule

This structure contains the definition for a Contributor Insights rule.

InsightRuleContributor

One of the unique contributors found by a Contributor Insights rule. If the rule contains multiple keys, then

InsightRuleContributorDatapoint

One data point related to one contributor.

InsightRuleMetricDatapoint

One data point from the metric time series returned in a Contributor Insights rule report.

LabelOptions

This structure includes the Timezone parameter, which you can use

MessageData

A message returned by the GetMetricDataAPI, including a code and a description.

Metric

Represents a specific metric.

MetricAlarm

The details about a metric alarm.

MetricDataQuery

This structure is used in both GetMetricData and PutMetricAlarm. The supported

MetricDataResult

A GetMetricData call returns an array of MetricDataResult

MetricDatum

Encapsulates the information sent to either create a metric or add new values

MetricStat

This structure defines the metric to be returned, along with the statistics, period, and units.

MetricStreamEntry

This structure contains the configuration information about one metric stream.

MetricStreamFilter

This structure contains the name of one of the metric namespaces that is listed in

PartialFailure

This array is empty if the API operation was successful for all the rules specified in the request. If the operation could

Range

Specifies one range of days or times to exclude from use for training an

StatisticSet

Represents a set of statistics that describes a specific metric.

Tag

A key-value pair associated with a CloudWatch resource.

Enums

AlarmType

AnomalyDetectorStateValue

ComparisonOperator

HistoryItemType

MetricStreamOutputFormat

RecentlyActive

ScanBy

StandardUnit

StateValue

Statistic

StatusCode