Module aws_sdk_cloudwatch::operation

Structs

DeleteAlarms

Deletes the specified alarms. You can delete up to 100 alarms in one operation. However, this total can include no more

DeleteAnomalyDetector

Deletes the specified anomaly detection model from your account.

DeleteDashboards

Deletes all dashboards that you specify. You

DeleteInsightRules

Permanently deletes the specified Contributor Insights rules.

DeleteMetricStream

Permanently deletes the metric stream that you specify.

DescribeAlarmHistory

Retrieves the history for the specified alarm. You can filter the results by date range or item type.

DescribeAlarms

Retrieves the specified alarms. You can filter the results by specifying a prefix for the alarm

DescribeAlarmsForMetric

Retrieves the alarms for the specified metric. To

DescribeAnomalyDetectors

Lists the anomaly detection models that you have created in your account. You can list all

DescribeInsightRules

Returns a list of all the Contributor Insights rules in your account.

DisableAlarmActions

Disables the actions for the specified alarms. When an alarm's actions are disabled, the

DisableInsightRules

Disables the specified Contributor Insights rules. When rules are disabled, they do not analyze log groups and do

EnableAlarmActions

Enables the actions for the specified alarms.

EnableInsightRules

Enables the specified Contributor Insights rules. When rules are enabled, they immediately begin analyzing log data.

GetDashboard

Displays the details of the dashboard that you specify.

GetInsightRuleReport

This operation returns the time series data collected by a Contributor Insights rule. The data includes the identity and number of

GetMetricData

You can use the GetMetricData API to retrieve as many as 500 different

GetMetricStatistics

Gets statistics for the specified metric.

GetMetricStream

Returns information about the metric stream that you specify.

GetMetricWidgetImage

You can use the GetMetricWidgetImage API to retrieve a snapshot graph of

ListDashboards

Returns a list of the dashboards for your account. If you include DashboardNamePrefix, only

ListMetricStreams

Returns a list of metric streams in this account.

ListMetrics

List the specified metrics. You can use the returned metrics with

ListTagsForResource

Displays the tags associated with a CloudWatch resource. Currently, alarms

PutAnomalyDetector

Creates an anomaly detection model for a CloudWatch metric. You can use the model

PutCompositeAlarm

Creates or updates a composite alarm. When you create a composite

PutDashboard

Creates a dashboard if it does not already exist, or updates an existing dashboard. If you update a dashboard,

PutInsightRule

Creates a Contributor Insights rule. Rules evaluate log events in a

PutMetricAlarm

Creates or updates an alarm and associates it with the specified metric, metric math expression,

PutMetricData

Publishes metric data points to Amazon CloudWatch. CloudWatch associates

PutMetricStream

Creates or updates a metric stream. Metric streams can automatically stream CloudWatch metrics

SetAlarmState

Temporarily sets the state of an alarm for testing purposes. When the updated

StartMetricStreams

Starts the streaming of metrics for one or more of your metric streams.

StopMetricStreams

Stops the streaming of metrics for one or more of your metric streams.

TagResource

Assigns one or more tags (key-value pairs) to the specified CloudWatch resource. Currently, the only CloudWatch resources that

UntagResource

Removes one or more tags from the specified resource.